+++ src/CMakeLists.txt
@@ -2,7 +2,7 @@
 add_library( pm SHARED element_influence.cpp panel3.cpp panel2.cpp case3.cpp case2.cpp  lifting_line.cpp)
 set_target_properties(pm PROPERTIES PREFIX "")
 
-PYTHON_ADD_MODULE(_ppm SHARED ppm_ext_1.cpp)
+PYTHON_ADD_MODULE(_ppm SHARED ppm_ext.cpp)
 set_target_properties(_ppm PROPERTIES LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/ppm)
 set_target_properties(_ppm PROPERTIES PREFIX "")
 target_link_libraries(_ppm ${PY_EIGEN} pm)
